data_pipeline_aws/producer/requirements.txt
kafka-python==2.0.2
faker==22.0.0
click==8.1.7
pydantic==2.5.3
python-dotenv==1.0.0
pyyaml==6.0.1
